+++ controllers/itemController.js
@@ -120,7 +120,31 @@ async function postUpdateForm(req, res) {
   }
 }
 
+async function getDeleteForm(req, res) {
+  try {
+    const itemResult = await db.getById(req.params.id, "items");
+    const item = itemResult[0];
+    res.render("deleteForm", { entity: item, path: "items", errors: [] });
+  } catch (err) {
+    console.error(err);
+  }
+}
+
 async function postDeleteItem(req, res) {
+  await validatePassword[0].run(req);
+
+  const errors = validationResult(req);
+
+  if (!errors.isEmpty()) {
+    const itemResult = await db.getById(req.params.id, "items");
+    const item = itemResult[0];
+    return res.render("deleteForm", {
+      entity: item,
+      path: "items",
+      errors: errors.array(),
+    });
+  }
+
   try {
     const id = req.params.id;
     await db.deleteById(id, "items");
@@ -137,4 +161,5 @@ module.exports = {
   getUpdateForm,
   postUpdateForm,
   postDeleteItem,
+  getDeleteForm,
 };
